backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
alembic
aiofiles
python-multipart
discord.py
python-dotenv
